+++ src/Firebase/ServiceAccount/Discoverer.php
@@ -9,6 +9,9 @@ use Kreait\Firebase\Exception\ServiceAccountDiscoveryFailed;
 use Kreait\Firebase\ServiceAccount;
 use Throwable;
 
+/**
+ * @deprecated 4.42
+ */
 class Discoverer
 {
     /**